tests/test_webhook.py
"""
Unit tests for WhatsApp webhook endpoints
"""

import pytest
from fastapi.testclient import TestClient
from unittest.mock import patch, AsyncMock

# These tests assume the main.py is importable
# Adjust imports based on your actual project structure

def test_health_check():
    """Test health check endpoint"""
    # This would require importing the app from main.py
    # Example: from main import app
    # client = TestClient(app)
    # response = client.get("/health")
    # assert response.status_code == 200
    # assert response.json()["status"] == "healthy"
    pass

def test_webhook_verification_valid_token():
    """Test webhook verification with valid token"""
    # Example test structure:
    # client = TestClient(app)
    # response = client.get(
    #     "/webhook",
    #     params={
    #         "hub.verify_token": "lexia-aoc-2602",
    #         "hub.challenge": "test_challenge_123"
    #     }
    # )
    # assert response.status_code == 200
    # assert response.text == "test_challenge_123"
    pass

def test_webhook_verification_invalid_token():
    """Test webhook verification with invalid token"""
    # Example test structure:
    # client = TestClient(app)
    # response = client.get(
    #     "/webhook",
    #     params={
    #         "hub.verify_token": "invalid_token",
    #         "hub.challenge": "test_challenge_123"
    #     }
    # )
    # assert response.status_code == 403
    pass

def test_webhook_message_processing():
    """Test webhook message processing"""
    # Example test structure:
    # client = TestClient(app)
    # payload = {
    #     "object": "whatsapp_business_account",
    #     "entry": [{
    #         "changes": [{
    #             "value": {
    #                 "messages": [{
    #                     "from": "551199999999",
    #                     "id": "wamid.test",
    #                     "type": "text",
    #                     "text": {"body": "Hello"}
    #                 }]
    #             }
    #         }]
    #     }]
    # }
    # response = client.post("/webhook", json=payload)
    # assert response.status_code == 200
    # assert response.json()["status"] == "ok"
    pass

def test_root_endpoint():
    """Test root endpoint"""
    # Example test structure:
    # client = TestClient(app)
    # response = client.get("/")
    # assert response.status_code == 200
    # data = response.json()
    # assert data["service"] == "Lexia WhatsApp Agent"
    # assert "endpoints" in data
    pass

if __name__ == "__main__":
    pytest.main([__file__, "-v"])
