requirements.txt
# Google ADK and dependencies
google-adk>=0.1.0
google-cloud-vertexai>=1.0.0
google-generativeai>=0.7.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP client
aiohttp>=3.9.0
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0

# Logging and monitoring
python-json-logger>=2.0.0

# Additional utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
